+++ responses.csv
@@ -15,3 +15,38 @@
 "0","0","0","0","0","0"
 "0","0","0","0","0","0"
 "15","0","0","4","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"0","0","0","0","0","0"
+"1","1","0","0","0","0"
+"0","0","0","0","0","0"
+"0","1","0","0","0","0"
+"1","0","0","0","0","0"
+"1","0","0","0","0","0"
+"1","0","0","0","0","0"
+"1","0","0","0","0","0"
+"3","0","0","0","0","0"
+"2","0","0","0","0","0"
