+++ weblm/main.py
@@ -12,7 +12,7 @@ from multiprocessing import Pool
 
 import cohere
 
-from .controller import Command, Controller, Prompt
+from .basic_controller.controller import Command, Controller, Prompt
 from .crawler import URL_PATTERN, Crawler
 
 co = cohere.Client(os.environ.get("COHERE_KEY"), check_api_key=False)
